separate attach logic from the rest of the catalog
